Feat/get receipts

+++ CaptureReceipt/src/main/java/com/mytiki/sdk/capture/receipt/capacitor/email/Email.kt
@@ -26,6 +26,7 @@ import com.mytiki.capture_receipt.email.getImapScanTime
 import com.mytiki.capture_receipt.email.setImapScanTime
 import com.mytiki.sdk.capture.receipt.capacitor.account.Account
 import com.mytiki.sdk.capture.receipt.capacitor.account.AccountCommon
+import com.mytiki.sdk.capture.receipt.capacitor.receipt.Receipt
 import kotlinx.coroutines.CompletableDeferred
 import kotlinx.coroutines.MainScope
 import kotlinx.coroutines.async
@@ -142,7 +143,7 @@ class Email(
      */
     fun scrape(
         context: Context,
-        onReceipt: (receipt: ScanResults?) -> Unit,
+        onReceipt: (receipt: Receipt?) -> Unit,
         onError: (msg: String) -> Unit,
         onComplete: () -> Unit
     ) {
@@ -161,33 +162,13 @@ class Email(
                         onComplete()
                     } else {
                         client.dayCutoff(dayCutOff)
-
-                        client.messages(object : MessagesCallback {
-
-                            @SuppressLint("SetTextI18n")
-                            override fun onComplete(
-                                credential: PasswordCredentials,
-                                result: List<ScanResults>
-                            ) {
-                                Timberland.d("credentials $credential results $result")
-
-                            }
-
-                            @SuppressLint("SetTextI18n")
-                            override fun onException(throwable: Throwable) {
-                                Timberland.e(throwable)
-
-                            }
-                        })
-
-
                        client.messages(object : MessagesCallback {
                            override fun onComplete(
                                credential: PasswordCredentials,
                                result: List<ScanResults>
                            ) {
                                result.forEach { receipt ->
-                                   onReceipt(receipt)
+                                   onReceipt(Receipt.opt(receipt))
                                }
                                context.setImapScanTime(now)
                                onComplete()
@@ -199,6 +180,49 @@ class Email(
                                client.close()
                            }
                        })
+                    }
+                }.addOnFailureListener {
+                    onComplete()
+                }
+            }
+        }
+    }
+
+    fun scrape(
+        context: Context,
+        account: Account,
+        onReceipt: (receipt: Receipt?) -> Unit,
+        onError: (String) -> Unit,
+        onComplete: () -> Unit
+    ) {
+        MainScope().async {
+            var dayCutOff = 15
+            val now = Calendar.getInstance().timeInMillis
+            val lastScrape = context.getImapScanTime().await()
+            val diffInMillis = now - lastScrape
+            val diffInDays = floor((diffInMillis / 86400000).toDouble()).toInt()
+            if (diffInDays <= 15) {
+                dayCutOff = diffInDays
+            }
+            client(context, onError) { client ->
+                client.accounts().addOnSuccessListener { credentials ->
+                    val emailAccount = credentials.firstOrNull {it.username() == account.username}
+                    if (emailAccount == null) {
+                        onComplete()
+                    } else {
+                        client.dayCutoff(dayCutOff)
+                        client.messages(emailAccount).addOnSuccessListener {result ->
+                            result.forEach { receipt ->
+                                onReceipt(Receipt.opt(receipt))
+                            }
+                            context.setImapScanTime(now)
+                            onComplete()
+                            client.close()
+                        }.addOnFailureListener {throwable ->
+                            onError(throwable.message.toString())
+                            onComplete()
+                            client.close()
+                        }
                     }
                 }.addOnFailureListener {
                     onComplete()
@@ -267,19 +291,23 @@ class Email(
     ) {
         this.client(context, onError) { client ->
             client.accounts().addOnSuccessListener { list ->
-                val passwordCredentials = list.first {
-                    it.username() == account.username && it.provider() == EmailEnum.fromString(
-                        account.accountCommon.id
-                    ).value
+                val passwordCredentials = list.firstOrNull {
+                    it.username() == account.username && it.provider() == EmailEnum.fromString(account.accountCommon.id).value
                 }
-                client.clearLastCheckedTime(Provider.valueOf(account.accountCommon.id))
-                context.deleteImapScanTime()
-                client.logout(passwordCredentials).addOnSuccessListener {
-                    onRemove()
-                }.addOnFailureListener {
+                if (passwordCredentials != null) {
+                    client.clearLastCheckedTime(Provider.valueOf(account.accountCommon.id))
+                    context.deleteImapScanTime()
+                    client.logout(passwordCredentials).addOnSuccessListener {
+                        onRemove()
+                    }.addOnFailureListener {
+                        onError(
+                            it.message
+                                ?: "Unknown error when removing account ${account.username} from ${account.accountCommon.id}"
+                        )
+                    }
+                } else {
                     onError(
-                        it.message
-                            ?: "Unknown error when removing account ${account.username} from ${account.accountCommon.id}"
+                        "Unknown error when removing account ${account.username} from ${account.accountCommon.id}"
                     )
                 }
             }
@@ -297,7 +325,7 @@ class Email(
     fun flush(context: Context, onComplete: () -> Unit, onError: (msg: String) -> Unit) {
         this.client(context, onError) { client ->
             client.clearLastCheckedTime()
-            //context.deleteImapScanTime()
+            context.deleteImapScanTime()
             client.logout().addOnSuccessListener {
                 onComplete()
             }.addOnFailureListener {
